Add modified `pkey_from_cryptography_key()`
It is a modified version of `OpenSSL.crypto.PKey.from_cryptography_key()`
which also accepts EC Keys.

Using this function, a patched PyOpenSSL is no longer required, which
makes it possiblel to remove the `--process-dependency-links` argument
at the pip-install.

There is a pull request for `OpenSSL.crypto.PKey.from_cryptography_key()` of
PyOpenSSL to also accept EC Keys, but there are comprehensible reasons not to
accept this request in the near future -- or at all:
https://github.com/pyca/pyopenssl/pull/636